move FormerContainer to own file
